Cleanup addonInstaller - caling std explicitly - fixed type mismatch - removed "worlds" from list of mod subfolders - RequestExecution() opens folder in explorer - replaced "sprintf" with "sprintf_s", "fopen" with "fopen_s"
gameRestart
- calling std explicitly
- fixed type mismatch
- replaced "strcmpi" with "_stricmp", "strncmpi" with "_strnicmp", "sprintf" with "sprintf_s", "fopen" with "fopen_s"
- removed queryserver option

fwatch.dll
- fixed type mismatch in fnv1a_hash()

+++ TestDLL/_Functions.cpp
@@ -1798,8 +1798,8 @@ int VerifyPath(String &path, StringDynamic &buffer, int options) {
 
 
 //https://stackoverflow.com/questions/11413860/best-string-hashing-function-for-short-filenames
-unsigned int fnv1a_hash(unsigned int hash, char *text, int text_length, bool lowercase) {
-    for (int i=0; i<text_length; i++)
+unsigned int fnv1a_hash(unsigned int hash, char *text, size_t text_length, bool lowercase) {
+    for (size_t i=0; i<text_length; i++)
 		hash = (hash ^ (lowercase ? tolower(text[i]) : text[i])) * FNV_PRIME;
 
     return hash;
